move plotting fxn to plot_from_npy.py for speed - importing keras takes a while
