+++ cmiyc/custom_vae.py
@@ -39,6 +39,7 @@ import time
 import dataset_utils
 import viz_utils
 from vanilla_vae import VanillaVae
+from plot_from_npy import plot_data
 
 class SplitReconKL(Callback):
 	'''
@@ -113,42 +114,6 @@ class SplitReconKL(Callback):
 		save_fn = self.imgfn
 
 		plot_data(data, save_dir, save_fn)
-
-def plot_data(data, save_dir, save_fn, title=None):
-	# Generate a x spanning the number of epochs
-		epoch_n = data.shape[1]
-		x = np.linspace(1, epoch_n, epoch_n)
-		y1 = data[0,:] # total
-		y2 = data[1,:] # recon
-		y3 = data[2,:] # KL
-		
-		plt.figure(figsize=(10,8))
-		plt.stackplot(x, y2, y3, labels=['Reconstruction term','KL term'])
-		
-		if title:
-			plt.title(title)
-		else:
-			plt.title("Training log-loss: Reconstruction vs KL term breakdown")
-		
-		plt.xlabel("Epoch")
-		if epoch_n < 20:
-			x_tick_interval = 1
-		elif epoch_n <= 100:
-			x_tick_interval = 5
-		else:
-			x_tick_interval = 10
-
-		plt.xticks(np.arange(min(x), max(x)+1, x_tick_interval))
-
-		plt.ylabel("Log-Loss")
-		plt.yscale("log")
-
-		plt.legend(loc='upper left')
-		if save_img:
-			plt.savefig(saved_dir+save_fn)
-			print("Saved {}".format(saved_dir+save_fn))
-		else:
-			plt.show()
 
 class CustomVae(VanillaVae):
 	'''
