+++ cmiyc/plot_from_npy.py
@@ -3,6 +3,42 @@ import numpy as np
 import matplotlib.pyplot as plt
 import seaborn as sns
 
+def plot_data(data, save_dir, save_fn, title=None):
+	# Generate a x spanning the number of epochs
+		epoch_n = data.shape[1]
+		x = np.linspace(1, epoch_n, epoch_n)
+		y1 = data[0,:] # total
+		y2 = data[1,:] # recon
+		y3 = data[2,:] # KL
+		
+		plt.figure(figsize=(10,8))
+		plt.stackplot(x, y2, y3, labels=['Reconstruction term','KL term'])
+		
+		if title:
+			plt.title(title)
+		else:
+			plt.title("Training log-loss: Reconstruction vs KL term breakdown")
+		
+		plt.xlabel("Epoch")
+		if epoch_n < 20:
+			x_tick_interval = 1
+		elif epoch_n <= 100:
+			x_tick_interval = 5
+		else:
+			x_tick_interval = 10
+
+		plt.xticks(np.arange(min(x), max(x)+1, x_tick_interval))
+
+		plt.ylabel("Log-Loss")
+		plt.yscale("log")
+
+		plt.legend(loc='upper left')
+		if save_img:
+			plt.savefig(saved_dir+save_fn)
+			print("Saved {}".format(saved_dir+save_fn))
+		else:
+			plt.show()
+
 def plot_from_npy():
 	'''
 	Using all .npy files spit out by train-all-sigs, generate a stacked area
@@ -21,6 +57,13 @@ def plot_from_npy():
 		fp = './saved-models/loss_splits/' + potential_fn
 		if os.path.isfile(fp):
 			data = np.load(fp)
+
+			# Convert values of col 1 and 2 to percentage of col 0,
+			# to normalize for different loss scales on different sigs
+			data[0] /= data[0]
+			data[1] /= data[0]
+			data[2] /= data[0] 
+
 			data_list.append(data)
 			print("File found for sig_id={}, shape is {}".format(sig_id, data.shape))
 
@@ -28,7 +71,10 @@ def plot_from_npy():
 	print("All sig_ids processed.")
 
 	all_data = np.array(data_list)
-	print(all_data.shape)
+	
+	# Take mean down the first axis
+	mean_data = np.mean(all_data, axis=0)
+	print(mean_data.shape)
 
 def main():
